+++ src/interfaces/prescription.interface.ts
@@ -18,7 +18,7 @@ export interface PrescriptionSupply {
 
 /* eslint no-extra-semi: "error"*/
 export default interface IPrescription extends Document {
-    prescriptionId?: number;
+    prescriptionId?: string;
     patient: IPatient;
     professional: {
         userId: string;
